% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diff_to_heatwach.R
\name{add_diff_heatwatch}
\alias{add_diff_heatwatch}
\title{Calculate difference to heatwatch product.}
\usage{
add_diff_heatwatch(products)
}
\arguments{
\item{products}{list. List of gridded products.}
}
\description{
Calculate difference to heatwatch product.
}
\author{
Eva Marques
}
