% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_covariates_to_analysis.R
\name{add_imp}
\alias{add_imp}
\title{Add imperviousness to gridded product.}
\usage{
add_imp(products, imp_path)
}
\arguments{
\item{products}{list. List of gridded products.}

\item{imp_path}{character. Path to imperviousness raster.}
}
\description{
Add imperviousness to gridded product.
}
\author{
Eva Marques
}
