% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_covariates_to_analysis.R
\name{add_nlcd}
\alias{add_nlcd}
\title{Add National Land Cover Classes to gridded product.}
\usage{
add_nlcd(products, nlcd_path)
}
\arguments{
\item{products}{list. List of gridded products.}

\item{nlcd_path}{character. Path to NLCD raster.}
}
\description{
Add National Land Cover Classes to gridded product.
}
\author{
Eva Marques
}
