% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diff_to_heatwach.R
\name{calc_all_diff}
\alias{calc_all_diff}
\title{Calculate difference to heatwatch product for every gridded products.}
\usage{
calc_all_diff(
  hw_dict,
  gridmet_dir,
  daymet_dir,
  moment_of_day,
  imp_path,
  nlcd_path,
  as_df = TRUE
)
}
\arguments{
\item{hw_dict}{list. Contains list of heatwatch cities.}

\item{gridmet_dir}{character. Path to gridmet data.}

\item{daymet_dir}{character. Path to daymet data.}

\item{moment_of_day}{character. "am" or "af".}

\item{imp_path}{character. Path to imperviousness data.}

\item{nlcd_path}{character. Path to National Land Cover Dataset.}

\item{as_df}{boolean. If TRUE, returns a data.frame.}
}
\value{
Differences calculated for all the products.
}
\description{
Calculate difference to heatwatch product for every gridded products.
}
\author{
Eva Marques
}
