% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ghcnd_station.R
\name{check_ref_period}
\alias{check_ref_period}
\title{Validate station for anomaly calculation}
\usage{
check_ref_period(site_id, var)
}
\arguments{
\item{site_id}{The station ID}

\item{var}{The variable to check}
}
\value{
A boolean indicating if the station is valid
}
\description{
Check if a station is valid for anomaly calculation
}
\author{
Eva Marques
}
