% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_temp.R
\name{convert_temp}
\alias{convert_temp}
\title{Convert temperature between Celsius, Fahrenheit, and Kelvin}
\usage{
convert_temp(x, from, to)
}
\arguments{
\item{x}{numeric vector}

\item{from}{character vector, one of "C", "F", or "K"}

\item{to}{character vector, one of "C", "F", or "K"}
}
\value{
numeric vector
}
\description{
Convert temperature between Celsius, Fahrenheit, and Kelvin
}
\author{
Eva Marques
}
