% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_heatwatch_city.R
\name{create_heatwatch_dictionary}
\alias{create_heatwatch_dictionary}
\title{Create heatwatch dictionary as a data.frame}
\usage{
create_heatwatch_dictionary(storage_dir)
}
\arguments{
\item{storage_dir}{The path to the heatwatch rasters and traverses shp}
}
\value{
A data frame with the heatwatch dictionary
}
\description{
Create heatwatch dictionary as a data.frame
}
\author{
Eva Marques
}
