% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_product.R
\name{crop_product}
\alias{crop_product}
\title{Crop a temperature product}
\usage{
crop_product(area_shp, product)
}
\arguments{
\item{area_shp}{a terra::SpatVector of a polygon}

\item{product}{a terra::SpatRaster or terra::SpatVector
of a temperature product}
}
\value{
the product cropped
}
\description{
Crop a temperature product
}
\author{
Eva Marques
}
