% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_daymet.R
\name{download_daymet_poly}
\alias{download_daymet_poly}
\title{Download Daymet data for entire US a given year}
\usage{
download_daymet_poly(storage_path, year, polygon, var)
}
\arguments{
\item{storage_path}{character path to store the data}

\item{year}{an integer}

\item{polygon}{a polygon}
}
\description{
Download Daymet data for entire US a given year
}
