% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ghcnd_station.R
\name{find_nearest_valid_ghcnd}
\alias{find_nearest_valid_ghcnd}
\title{Find nearest valid GHCND station to a given location}
\usage{
find_nearest_valid_ghcnd(lat, lon)
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longitude}
}
\value{
A list with the nearest station for TMAX, TMIN and TAVG
}
\description{
Find nearest valid GHCND station to a given location
}
\author{
Eva Marques
}
