% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGDELT2.R
\name{GetGDELT}
\alias{GetGDELT}
\title{ETL Egypt GDELT data}
\usage{
GetGDELT(DaysBack = NULL, writeResults = T,
  location = "~/GitHub/GiveMercuryAMoon/Data", Engine = "Egypt")
}
\arguments{
\item{DaysBack}{}

\item{location}{}
}
\value{
Returns a data frame to the global environment showing the debit/credit results.
}
\description{
# a giant for loop that downloads gdelt data for a given month, extracts the egypt related info, combines it, saves it to a master folder, and deletes them if all of the above is done successfully
}
\examples{
DebitCreditCompare(MasterFile)
}
