% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ghcnd_station.R
\name{list_valid_stations}
\alias{list_valid_stations}
\title{List valid stations}
\usage{
list_valid_stations(var)
}
\arguments{
\item{var}{The variable to check}
}
\value{
A data frame with the list of valid stations
}
\description{
List all valid stations for anomaly calculation
}
\author{
Eva Marques
}
