% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_city_borders.R
\name{load_city_borders}
\alias{load_city_borders}
\title{Load city borders}
\usage{
load_city_borders(city, us_cities_shp)
}
\arguments{
\item{city}{character. Name of the city}

\item{us_cities_shp}{character. Path to the US cities shapefile}
}
\value{
a vector with the city borders
}
\description{
Load city borders
}
\author{
Eva Marques
}
