% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_daymet.R
\name{load_daymet}
\alias{load_daymet}
\title{Load daymet data for a given date and area from storage folder
Daymet yearly data is downloaded from here
https://daac.ornl.gov/cgi-bin/dsviewer.pl?ds_id=2129
and file name is not changed.}
\usage{
load_daymet(dates, area, var, storage_folder)
}
\arguments{
\item{dates}{a vector of dates}

\item{area}{a polygon}

\item{var}{a character}

\item{storage_folder}{a character}
}
\value{
a raster
}
\description{
Load daymet data for a given date and area from storage folder
Daymet yearly data is downloaded from here
https://daac.ornl.gov/cgi-bin/dsviewer.pl?ds_id=2129
and file name is not changed.
}
\author{
Eva Marques
}
