% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gridmet.R
\name{load_gridmet_day}
\alias{load_gridmet_day}
\title{Load GridMET variable for a day}
\usage{
load_gridmet_day(day, gridmet_dir, var)
}
\arguments{
\item{day}{a date object}

\item{gridmet_dir}{the directory where GRIDMET data is stored}

\item{var}{the variable to load (accepted value: "tmmn", "tmmx")}
}
\value{
a raster object
}
\description{
Load GridMET variable for a day
}
