% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_heatwatch_city.R
\name{load_heatwatch_city}
\alias{load_heatwatch_city}
\title{Load the heatwatch data for a given city}
\usage{
load_heatwatch_city(city, var, heatwatch_dict)
}
\arguments{
\item{city}{The city to load the data for}

\item{var}{The variable to load ("am", "af" or "pm")}

\item{heatwatch_dict}{The path to the heatwatch dictionary file}
}
\value{
A raster with the heatwatch data for the city
}
\description{
Load the heatwatch data for a given city
}
\author{
Eva Marques
}
