% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_temp_products_city.R
\name{load_temp_products_city}
\alias{load_temp_products_city}
\title{Return air temperature products from 3 different sources for a city:
gridmet, daymet, and heatwatch}
\usage{
load_temp_products_city(city, hw_dict, gridmet_dir, daymet_dir, moment_of_day)
}
\arguments{
\item{city}{a character}

\item{hw_dict}{a data.frame}

\item{gridmet_dir}{a character}

\item{daymet_dir}{a character}

\item{moment_of_day}{a character: "am" or "af"}
}
\description{
Return air temperature products from 3 different sources for a city:
gridmet, daymet, and heatwatch
}
