% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_temp_heatwatch_cws.R
\name{map_cws_hw}
\alias{map_cws_hw}
\title{map_observations_hw}
\usage{
map_cws_hw(cws, hw_r_am, hw_r_af)
}
\arguments{
\item{cws}{A data frame with the observations from CWS.}

\item{hw_r_am}{A raster with the Heatwatch data for the morning (6-7am).}

\item{hw_r_af}{A raster with the Heatwatch data for the afternoon (3-4pm).}
}
\value{
A list with two ggplot objects.
}
\description{
This function creates two maps with the observations
from Heatwatch and CWS (minimum and maximum temperatures).
}
