% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_temp_daymet.R
\name{map_temp_daymet}
\alias{map_temp_daymet}
\title{Plot Daymet air temperature}
\usage{
map_temp_daymet(r, borders = NULL, temp_unit)
}
\arguments{
\item{r}{A raster object}

\item{borders}{A polygon}

\item{temp_unit}{A character string (either "C" or "F") indicating
the unit of r}
}
\value{
A ggplot2 object
}
\description{
Plot Daymet air temperature
}
