% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_temp_heatwatch.R
\name{map_temp_heatwatch}
\alias{map_temp_heatwatch}
\title{Plot heatwatch air temperature}
\usage{
map_temp_heatwatch(r, tr, borders = NULL, temp_unit)
}
\arguments{
\item{r}{A spatraster of heatwatch air temperature}

\item{tr}{A spatvector of heatwatch transects}

\item{borders}{A polygon}

\item{temp_unit}{A character string (either "C" or "F") indicating
the unit of r}
}
\value{
A ggplot2 object
}
\description{
Plot heatwatch air temperature
}
