% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_temp_city.R
\name{map_temp_products_city}
\alias{map_temp_products_city}
\title{map_temp_products_city}
\usage{
map_temp_products_city(city, hw_dict, gridmet_dir, daymet_dir, moment_of_day)
}
\arguments{
\item{city}{character. City name}

\item{hw_dict}{data.frame. Heatwatch dictionary}

\item{gridmet_dir}{character. Path to the GRIDmet data}

\item{daymet_dir}{character. Path to the Daymet data}

\item{moment_of_day}{character. Moment of the day (am or af)}
}
\value{
ggplot object
}
\description{
Create a map with the temperature products for a city
}
