% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercury.R
\name{mercury}
\alias{mercury}
\title{Mercury}
\usage{
mercury(urls)
}
\arguments{
\item{urls}{list of urls to parse with Mercury Web Parser API}
}
\value{
a tibble with many informations about the pages including title, author,
date_published, etc.
}
\description{
Mercury
}
\examples{
mercury("https://trackchanges.postlight.com/building-awesome-cms-f034344d8ed")

}
