% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercury.R
\name{mercury_key}
\alias{mercury_key}
\title{Function used to get the Mercury API Key.}
\usage{
mercury_key()
}
\description{
You need to get yours at [Mercury website](https://mercury.postlight.com/web-parser/)
}
