% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ghcnd_station.R
\name{open_station}
\alias{open_station}
\title{Open and format GHCNd station}
\usage{
open_station(site_id)
}
\arguments{
\item{site_id}{character. ID of the station.}
}
\value{
data.frame of GHCNh station data.
}
\description{
Open and format GHCNd station
}
\author{
Eva Marques
}
