% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temp_difference_to_heatwatch.R
\name{scatterplot_deltat_vs_imp}
\alias{scatterplot_deltat_vs_imp}
\title{scatterplot_deltat_vs_imp}
\usage{
scatterplot_deltat_vs_imp(diff_df_am, diff_df_af)
}
\arguments{
\item{diff_df_am}{data.frame. Data frame with the temperature difference to
heatwatch in the morning (6-7am)}

\item{diff_df_af}{data.frame. Data frame with the temperature difference to
heatwatch in the afternoon (3-4pm)}
}
\value{
ggplot objects
}
\description{
Plot temperature difference to HeatWatch in fonction of
imperviousness
}
