% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_heatwatch_city.R
\name{load_heatwatch_city}
\alias{load_heatwatch_city}
\title{Open heatwatch rasters}
\usage{
load_heatwatch_city(hw_dict, city)
}
\arguments{
\item{hw_dict}{list. Contains list of heatwatch metadata.}

\item{city}{character. Name of the city.}
}
\value{
all heatwatch data for the given city.
}
\description{
Open heatwatch rasters
}
\author{
Eva Marques
}
